requirements.txt
# requirements.txt
pytest==8.0.2
selenium==4.18.1
webdriver-manager==4.0.1
allure-pytest==2.13.2
pytest-xdist==3.5.0
faker==23.3.0
pytest-html==4.1.1
openpyxl==3.1.2